data-engineering/zoomcamp/assignments/final-project/data-generator/requirements.txt
faker
pandas
pyarrow
google-cloud-bigquery
google-cloud-storage
